improvement: adding authentication to some routes

+++ src/routes/deleteListing.js
@@ -1,3 +1,4 @@
+import * as admin from 'firebase-admin';
 import { db } from "../database";
 
 export const deleteListingRoute = {
@@ -5,9 +6,13 @@ export const deleteListingRoute = {
   path: '/api/listings/{id}',
   handler: async (req, h) => {
     const { id } = req.params;
+    const token = req.headers.authtoken;
+    const user = await admin.auth().verifyIdToken(token);
+    const userId = user.uid;
+
     await db.query(
-      'DELETE FROM listings WHERE id=?',
-      [id]
+      'DELETE FROM listings WHERE id=? and user_id=?',
+      [id, userId]
     );
 
     return { message: 'Success!' };
